refactor: GameSetting 상수 패키지 분리 - config 패키지로 분리
