+++ src/main/java/racingcar/config/GameSetting.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.config;
 
 public class GameSetting {
